Move components tests to a separate package
